third_party/WebKit/LayoutTests/external/wpt/cookies/samesite/window-open-expected.txt

This is a testharness.js-based test.
PASS Same-host auxiliary navigations are strictly same-site
PASS Subdomain auxiliary navigations are strictly same-site
FAIL Cross-site auxiliary navigations are laxly same-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.4397358922158694" but got (undefined) undefined
PASS Same-host redirecting to same-host auxiliary navigations are strictly same-site
PASS Subdomain redirecting to same-host auxiliary navigations are strictly same-site
PASS Cross-site redirecting to same-host auxiliary navigations are strictly same-site
PASS Same-host redirecting to subdomain auxiliary navigations are strictly same-site
PASS Subdomain redirecting to subdomain auxiliary navigations are strictly same-site
PASS Cross-site redirecting to subdomain auxiliary navigations are strictly same-site
FAIL Same-host redirecting to cross-site auxiliary navigations are laxly same-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.5391865782285838" but got (undefined) undefined
FAIL Subdomain redirecting to cross-site auxiliary navigations are laxly same-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.5210046826987287" but got (undefined) undefined
FAIL Cross-site redirecting to cross-site auxiliary navigations are laxly same-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.31798167929837184" but got (undefined) undefined
Harness: the test ran to completion.
